fixes Statfs_t syscalls for macOs

+++ go/database/mpt/tool/stress.go
@@ -361,7 +361,7 @@ func getFreeSpace(path string) (int64, error) {
 	if err != nil {
 		return 0, err
 	}
-	return int64(fs.Bavail) * fs.Bsize, nil
+	return int64(fs.Bavail) * int64(fs.Bsize), nil
 }
 
 func getMemoryUsage() uint64 {
